Feat (handler): add feature to find a handler for template resources
