+++ src/main/java/webserver/Handlers.java
@@ -2,6 +2,9 @@ package webserver;
 
 import http.request.Request;
 import webserver.customhandler.StaticResourceHandler;
+import webserver.customhandler.TemplateHandler;
+import webserver.resources.StaticResources;
+import webserver.resources.TemplateResources;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -18,6 +21,10 @@ public class Handlers {
 
         if (StaticResources.match(url)) {
             return new StaticResourceHandler();
+        }
+
+        if (TemplateResources.match(url)) {
+            return new TemplateHandler();
         }
 
         return handlers.stream()
